DRAW-350 fix: Package 수정
