+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/WebSocketRequest.kt
@@ -1,0 +1,6 @@
+package com.xorker.draw.websocket.message.request
+
+data class WebSocketRequest(
+    val action: RequestAction,
+    val body: String?,
+)
