--- app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/dto/WebSocketRequest.kt
@@ -1,8 +1,0 @@
-package com.xorker.draw.websocket.message.request.dto
-
-import com.xorker.draw.websocket.message.request.RequestAction
-
-data class WebSocketRequest(
-    val action: RequestAction,
-    val body: String?,
-)
