+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/mafia/MafiaGameRandomMatchingRequest.kt
@@ -1,4 +1,4 @@
-package com.xorker.draw.websocket.message.request.dto.game
+package com.xorker.draw.websocket.message.request.mafia
 
 data class MafiaGameRandomMatchingRequest(
     val accessToken: String,
