+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/request/mafia/MafiaGameReactionRequest.kt
@@ -1,4 +1,4 @@
-package com.xorker.draw.websocket.message.request.dto.game
+package com.xorker.draw.websocket.message.request.mafia
 
 data class MafiaGameReactionRequest(
     val reaction: String,
