+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/MafiaGameMessengerImpl.kt
@@ -9,20 +9,20 @@ import com.xorker.draw.mafia.assertIs
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.broker.WebSocketBroadcaster
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameAnswerBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameAnswerMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameDrawMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGamePlayerListBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGamePlayerListMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameReactionBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameReactionMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameTurnInfoBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameTurnInfoMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameVoteStatusBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaGameVoteStatusMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaRandomMatchingBody
-import com.xorker.draw.websocket.message.response.dto.game.MafiaRandomMatchingMessage
-import com.xorker.draw.websocket.message.response.dto.game.toResponse
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameAnswerBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameAnswerMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameDrawMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaGamePlayerListBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGamePlayerListMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameReactionBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameReactionMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameTurnInfoBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameTurnInfoMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameVoteStatusBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameVoteStatusMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaRandomMatchingBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaRandomMatchingMessage
+import com.xorker.draw.websocket.message.response.mafia.toResponse
 import org.springframework.stereotype.Component
 
 @Component
