+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/MafiaPhaseMessengerImpl.kt
@@ -9,22 +9,21 @@ import com.xorker.draw.mafia.MafiaPlayer
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.broker.WebSocketBroadcaster
-import com.xorker.draw.websocket.message.response.dto.game.toResponse
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaGameInfoBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaGameInfoMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseEndBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseEndMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseInferAnswerBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseInferAnswerMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhasePlayingBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhasePlayingMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseReadyBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseReadyMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseVoteBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseVoteMessage
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseWaitBody
-import com.xorker.draw.websocket.message.response.dto.phase.MafiaPhaseWaitMessage
-import com.xorker.draw.websocket.message.response.dto.phase.toResponse
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameInfoBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaGameInfoMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseEndBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseEndMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseInferAnswerBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseInferAnswerMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhasePlayingBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhasePlayingMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseReadyBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseReadyMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseVoteBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseVoteMessage
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseWaitBody
+import com.xorker.draw.websocket.message.response.mafia.MafiaPhaseWaitMessage
+import com.xorker.draw.websocket.message.response.mafia.toResponse
 import org.springframework.stereotype.Component
 
 @Component
