+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/mafia/MafiaPhaseEndMessage.kt
@@ -1,19 +1,21 @@
-package com.xorker.draw.websocket.message.response.dto.phase
+package com.xorker.draw.websocket.message.response.mafia
 
 import com.xorker.draw.websocket.message.response.ResponseAction
 import com.xorker.draw.websocket.message.response.SessionMessage
 import java.time.LocalDateTime
 
-data class MafiaPhaseInferAnswerMessage(
-    override val body: MafiaPhaseInferAnswerBody,
+data class MafiaPhaseEndMessage(
+    override val body: MafiaPhaseEndBody,
 ) : SessionMessage {
-    override val action = ResponseAction.PHASE_INFER_ANSWER
+    override val action = ResponseAction.PHASE_END
     override val status = SessionMessage.Status.OK
 }
 
-data class MafiaPhaseInferAnswerBody(
+data class MafiaPhaseEndBody(
     val startTime: LocalDateTime,
     val mafiaGameInfo: MafiaGameInfoMessage? = null,
+    val showAnswer: Boolean,
     val mafiaAnswer: String?,
+    val isMafiaWin: Boolean,
     val draw: List<Map<String, Any>>,
 )
