+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/mafia/MafiaPhaseWaitMessage.kt
@@ -1,9 +1,8 @@
-package com.xorker.draw.websocket.message.response.dto.phase
+package com.xorker.draw.websocket.message.response.mafia
 
 import com.xorker.draw.room.RoomId
 import com.xorker.draw.websocket.message.response.ResponseAction
 import com.xorker.draw.websocket.message.response.SessionMessage
-import com.xorker.draw.websocket.message.response.dto.game.MafiaPlayerResponse
 
 data class MafiaPhaseWaitMessage(
     override val body: MafiaPhaseWaitBody,
