+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/session/SessionFactory.kt
@@ -8,10 +8,8 @@ import com.xorker.draw.user.User
 import com.xorker.draw.user.UserId
 import com.xorker.draw.websocket.Session
 import com.xorker.draw.websocket.WaitingQueueSession
-import com.xorker.draw.websocket.message.request.SessionWrapper
-import com.xorker.draw.websocket.message.request.WaitingQueueSessionWrapper
-import com.xorker.draw.websocket.message.request.dto.game.MafiaGameRandomMatchingRequest
-import com.xorker.draw.websocket.message.request.dto.game.SessionInitializeRequest
+import com.xorker.draw.websocket.message.request.mafia.MafiaGameRandomMatchingRequest
+import com.xorker.draw.websocket.message.request.mafia.SessionInitializeRequest
 import org.slf4j.MDC
 import org.springframework.stereotype.Component
 import org.springframework.web.socket.WebSocketSession
